Removed dropdown within a navbar on mobile,
instead added the 2 links inside the dropdown

(register as donor and register as an organization)

directly in the ul menu on small devices.

On bigger screens the 2 links appear in the navbar instead

of the Register Now dropdown.

Also removed showDropdown state and handleDropdownToggle func
as neither is used now.